requirements.txt
Flask==2.3.2
PyPDF2==1.26.0
python-docx==0.8.11
Pillow==9.2.0
requests==2.28.1